railway-crossing/services/simple_sensor/requirements.txt
nats-py
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
protobuf